some refactor fileNames
